+++ src/upgrades/v2_1_0.py
@@ -1,10 +1,6 @@
 from db import genconn
+from app import config
 import traceback
-
-def upgrade_log(msg):
-    open(f"upgrades/log_v2_1_0.txt", "a").write(msg + "\n")
-    if not "WARN" in msg:
-        print(msg)
 
 def convert_quotation(s):
     s = str(s)
@@ -25,12 +21,12 @@ def run():
     conn = genconn(autocommit = True)
     cur = conn.cursor()
 
-    upgrade_log(f"Updating user table (reorder, add uid column)...")
+    print(f"Updating user table (reorder, add uid column)...")
     try:
         cur.execute(f"ALTER TABLE user RENAME TO user_old")
     except Exception as exc:
         if str(exc).find("already exists") != -1:
-            upgrade_log(f"Last upgrade seems to be failed, fixing up...")
+            print(f"Last upgrade seems to be failed, fixing up...")
             cur.execute(f"DROP TABLE user")
     cur.execute(f"SELECT * FROM user_old WHERE userid >= 0 ORDER BY userid ASC")
     rows1 = cur.fetchall()
@@ -60,14 +56,14 @@ def run():
         discordid2uid[row[1]] = uid
         if row[0] >= 0:
             userid2uid[row[0]] = uid
-    upgrade_log(f"Created {len(list(discordid2uid.keys()))} discordid -> uid links.")
+    print(f"Created {len(list(discordid2uid.keys()))} discordid -> uid links.")
     
-    upgrade_log(f"Updating user_password table (discordid -> uid)...")
+    print(f"Updating user_password table (discordid -> uid)...")
     try:
         cur.execute(f"ALTER TABLE user_password RENAME TO user_password_old")
     except Exception as exc:
         if str(exc).find("already exists") != -1:
-            upgrade_log(f"Last upgrade seems to be failed, fixing up...")
+            print(f"Last upgrade seems to be failed, fixing up...")
             cur.execute(f"DROP TABLE user_password")
     cur.execute(f"SELECT * FROM user_password_old")
     rows = cur.fetchall()
@@ -75,19 +71,19 @@ def run():
     for row in rows:
         row = process_row(row)
         if not row[0] in discordid2uid.keys():
-            upgrade_log(f"[WARN] Skipped {row[0]}: User does not exist.")
+            # print(f"[WARN] Skipped {row[0]}: User does not exist.")
             continue
         try:
             cur.execute(f"INSERT INTO user_password VALUES ({discordid2uid[row[0]]}, '{row[1]}', '{row[2]}')")
         except:
             traceback.print_exc()
     
-    upgrade_log(f"Updating user_activity table (discordid -> uid)...")
+    print(f"Updating user_activity table (discordid -> uid)...")
     try:
         cur.execute(f"ALTER TABLE user_activity RENAME TO user_activity_old")
     except Exception as exc:
         if str(exc).find("already exists") != -1:
-            upgrade_log(f"Last upgrade seems to be failed, fixing up...")
+            print(f"Last upgrade seems to be failed, fixing up...")
             cur.execute(f"DROP TABLE user_activity")
     cur.execute(f"SELECT * FROM user_activity_old")
     rows = cur.fetchall()
@@ -95,19 +91,19 @@ def run():
     for row in rows:
         row = process_row(row)
         if not row[0] in discordid2uid.keys():
-            upgrade_log(f"[WARN] Skipped {row[0]}: User does not exist.")
+            # print(f"[WARN] Skipped {row[0]}: User does not exist.")
             continue
         try:
             cur.execute(f"INSERT INTO user_activity VALUES ({discordid2uid[row[0]]}, '{row[1]}', {row[2]})")
         except:
             traceback.print_exc()
     
-    upgrade_log(f"Updating user_notification table (discordid -> uid)...")
+    print(f"Updating user_notification table (discordid -> uid)...")
     try:
         cur.execute(f"ALTER TABLE user_notification RENAME TO user_notification_old")
     except Exception as exc:
         if str(exc).find("already exists") != -1:
-            upgrade_log(f"Last upgrade seems to be failed, fixing up...")
+            print(f"Last upgrade seems to be failed, fixing up...")
             cur.execute(f"DROP TABLE user_notification")
     cur.execute(f"SELECT * FROM user_notification_old")
     rows = cur.fetchall()
@@ -115,19 +111,19 @@ def run():
     for row in rows:
         row = process_row(row)
         if not row[1] in discordid2uid.keys():
-            upgrade_log(f"[WARN] Skipped {row[1]}: User does not exist.")
+            # print(f"[WARN] Skipped {row[1]}: User does not exist.")
             continue
         try:
             cur.execute(f"INSERT INTO user_notification VALUES ({row[0]}, {discordid2uid[row[1]]}, '{row[2]}', {row[3]}, {row[4]})")
         except:
             traceback.print_exc()
     
-    upgrade_log(f"Updating banned table (discordid -> uid)...")
+    print(f"Updating banned table (discordid -> uid)...")
     try:
         cur.execute(f"ALTER TABLE banned RENAME TO banned_old")
     except Exception as exc:
         if str(exc).find("already exists") != -1:
-            upgrade_log(f"Last upgrade seems to be failed, fixing up...")
+            print(f"Last upgrade seems to be failed, fixing up...")
             cur.execute(f"DROP TABLE banned")
     cur.execute(f"SELECT * FROM banned_old")
     rows = cur.fetchall()
@@ -135,7 +131,7 @@ def run():
     for row in rows:
         row = process_row(row)
         if not row[0] in discordid2uid.keys():
-            upgrade_log(f"[WARN] Skipped {row[0]}: User does not exist.")
+            # print(f"[WARN] Skipped {row[0]}: User does not exist.")
             continue
         email = userinfo[row[0]][5]
         steamid = userinfo[row[0]][7]
@@ -145,32 +141,32 @@ def run():
         except:
             traceback.print_exc()
     
-    upgrade_log(f"Updating application table (discordid -> uid)...")
+    print(f"Updating application table (discordid -> uid)...")
     try:
         cur.execute(f"ALTER TABLE application RENAME TO application_old")
     except Exception as exc:
         if str(exc).find("already exists") != -1:
-            upgrade_log(f"Last upgrade seems to be failed, fixing up...")
+            print(f"Last upgrade seems to be failed, fixing up...")
             cur.execute(f"DROP TABLE application")
     cur.execute(f"SELECT * FROM application_old")
     rows = cur.fetchall()
-    cur.execute(f"CREATE TABLE application (applicationid INT AUTO_INCREMENT PRIMARY KEY, application_type INT, uid INT, data TEXT, status INT, submit_timestamp BIGINT, update_staff_userid INT, update_staff_timestamp BIGINT)")
+    cur.execute(f"CREATE TABLE application (applicationid INT AUTO_INCREMENT PRIMARY KEY, application_type INT, uid INT, data TEXT, status INT, submit_timestamp BIGINT, update_staff_userid INT, update_staff_timestamp BIGINT) DATA DIRECTORY = '{config.mysql_ext}'")
     for row in rows:
         row = process_row(row)
         if not row[2] in discordid2uid.keys():
-            upgrade_log(f"[WARN] Skipped {row[2]}: User does not exist.")
+            # print(f"[WARN] Skipped {row[2]}: User does not exist.")
             continue
         try:
             cur.execute(f"INSERT INTO application(applicationid, application_type, uid, data, status, submit_timestamp, update_staff_userid, update_staff_timestamp) VALUES ({row[0]}, {row[1]}, {discordid2uid[row[2]]}, '{row[3]}', {row[4]}, {row[5]}, {row[6]}, {row[5]})")
         except:
             traceback.print_exc()
     
-    upgrade_log(f"Updating session table (discordid -> uid)...")
+    print(f"Updating session table (discordid -> uid)...")
     try:
         cur.execute(f"ALTER TABLE session RENAME TO session_old")
     except Exception as exc:
         if str(exc).find("already exists") != -1:
-            upgrade_log(f"Last upgrade seems to be failed, fixing up...")
+            print(f"Last upgrade seems to be failed, fixing up...")
             cur.execute(f"DROP TABLE session")
     cur.execute(f"SELECT * FROM session_old")
     rows = cur.fetchall()
@@ -178,19 +174,19 @@ def run():
     for row in rows:
         row = process_row(row)
         if not row[1] in discordid2uid.keys():
-            upgrade_log(f"[WARN] Skipped {row[1]}: User does not exist.")
+            # print(f"[WARN] Skipped {row[1]}: User does not exist.")
             continue
         try:
             cur.execute(f"INSERT INTO session VALUES ('{row[0]}', {discordid2uid[row[1]]}, {row[2]}, '{row[3]}', '{row[4]}', '{row[5]}', {row[6]})")
         except:
             traceback.print_exc()
 
-    upgrade_log(f"Updating auth_ticket table (discordid -> uid)...")
+    print(f"Updating auth_ticket table (discordid -> uid)...")
     try:
         cur.execute(f"ALTER TABLE temp_identity_proof RENAME TO temp_identity_proof_old")
     except Exception as exc:
         if str(exc).find("already exists") != -1:
-            upgrade_log(f"Last upgrade seems to be failed, fixing up...")
+            print(f"Last upgrade seems to be failed, fixing up...")
             cur.execute(f"DROP TABLE auth_ticket")
     cur.execute(f"SELECT * FROM temp_identity_proof_old")
     rows = cur.fetchall()
@@ -202,19 +198,19 @@ def run():
     for row in rows:
         row = process_row(row)
         if not row[1] in discordid2uid.keys():
-            upgrade_log(f"[WARN] Skipped {row[1]}: User does not exist.")
+            # print(f"[WARN] Skipped {row[1]}: User does not exist.")
             continue
         try:
             cur.execute(f"INSERT INTO auth_ticket VALUES ('{row[0]}', {discordid2uid[row[1]]}, {row[2]})")
         except:
             traceback.print_exc()
 
-    upgrade_log(f"Updating application_token table (discordid -> uid)...")
+    print(f"Updating application_token table (discordid -> uid)...")
     try:
         cur.execute(f"ALTER TABLE application_token RENAME TO application_token_old")
     except Exception as exc:
         if str(exc).find("already exists") != -1:
-            upgrade_log(f"Last upgrade seems to be failed, fixing up...")
+            print(f"Last upgrade seems to be failed, fixing up...")
             cur.execute(f"DROP TABLE application_token")
     cur.execute(f"SELECT * FROM application_token_old")
     rows = cur.fetchall()
@@ -222,19 +218,19 @@ def run():
     for row in rows:
         row = process_row(row)
         if not row[2] in discordid2uid.keys():
-            upgrade_log(f"[WARN] Skipped {row[2]}: User does not exist.")
+            # print(f"[WARN] Skipped {row[2]}: User does not exist.")
             continue
         try:
             cur.execute(f"INSERT INTO application_token VALUES ('{row[0]}', '{row[1]}', {discordid2uid[row[2]]}, {row[3]}, {row[4]})")
         except:
             traceback.print_exc()
 
-    upgrade_log(f"Updating auditlog table (userid -> uid)...")
+    print(f"Updating auditlog table (userid -> uid)...")
     try:
         cur.execute(f"ALTER TABLE auditlog RENAME TO auditlog_old")
     except Exception as exc:
         if str(exc).find("already exists") != -1:
-            upgrade_log(f"Last upgrade seems to be failed, fixing up...")
+            print(f"Last upgrade seems to be failed, fixing up...")
             cur.execute(f"DROP TABLE auditlog")
     cur.execute(f"SELECT * FROM auditlog_old")
     rows = cur.fetchall()
@@ -242,19 +238,19 @@ def run():
     for row in rows:
         row = process_row(row)
         if not row[0] in userid2uid.keys():
-            upgrade_log(f"[WARN] Skipped {row[0]}: User does not exist.")
+            # print(f"[WARN] Skipped {row[0]}: User does not exist.")
             continue
         try:
             cur.execute(f"INSERT INTO auditlog VALUES ({userid2uid[row[0]]}, '{row[1]}', {row[2]})")
         except:
             traceback.print_exc()
     
-    upgrade_log(f"Updating settings table (discordid -> uid)...")
+    print(f"Updating settings table (discordid -> uid)...")
     try:
         cur.execute(f"ALTER TABLE settings RENAME TO settings_old")
     except Exception as exc:
         if str(exc).find("already exists") != -1:
-            upgrade_log(f"Last upgrade seems to be failed, fixing up...")
+            print(f"Last upgrade seems to be failed, fixing up...")
             cur.execute(f"DROP TABLE settings")
     cur.execute(f"SELECT * FROM settings_old")
     rows = cur.fetchall()
@@ -267,15 +263,15 @@ def run():
                 traceback.print_exc()
         else:
             if not row[0] in discordid2uid.keys():
-                upgrade_log(f"[WARN] Skipped {row[0]}: User does not exist.")
+                # print(f"[WARN] Skipped {row[0]}: User does not exist.")
                 continue
             try:
                 cur.execute(f"INSERT INTO settings VALUES ({discordid2uid[row[0]]}, '{row[1]}', '{row[2]}')")
             except:
                 traceback.print_exc()
     
-    skey2table = {"nxtuserid": "user", "nxtnotificationid": "user_notification", "nxtlogid": "dlog", "nxtappid": "application", "nxtannid": "announcement", "nxtchallengeid": "challenge", "nxtdownloadsid": "downloads", "nxteventid": "event"}
-    upgrade_log(f"Updating tables (add AUTO_INCREMENT property)")
+    skey2table = {"nxtnotificationid": "user_notification", "nxtlogid": "dlog", "nxtappid": "application", "nxtannid": "announcement", "nxtchallengeid": "challenge", "nxtdownloadsid": "downloads", "nxteventid": "event"}
+    print(f"Updating tables (add AUTO_INCREMENT property)")
     for skey in skey2table.keys():
         table = skey2table[skey]
         if table in ["user", "user_notification"]:
@@ -297,11 +293,11 @@ def run():
                 pass
             cur.execute(f"ALTER IGNORE TABLE {table} MODIFY COLUMN {idcolumn} INT NOT NULL AUTO_INCREMENT PRIMARY KEY")
 
-    upgrade_log(f"Updating settings table (nxtid)...")
+    print(f"Updating settings table (nxtid)...")
     for skey in skey2table.keys():
         cur.execute(f"SELECT sval FROM settings WHERE skey = '{skey}'")
         sval = cur.fetchone()[0]
         cur.execute(f"ALTER TABLE {skey2table[skey]} AUTO_INCREMENT={sval};")
         cur.execute(f"DELETE FROM settings WHERE skey = '{skey}'")
     
-    upgrade_log(f"Upgrade finished")
+    print(f"Upgrade finished")